backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]==3.1.19
celery==5.4.0
redis==5.0.7
pytest==8.2.0
pytest-django==4.9.0
reportlab